# Conflicts:
#	README.md
#	config.json.example
#	web

+++ pokemongo_bot/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+import os
 import logging
 import googlemaps
 import json
@@ -39,9 +40,9 @@ class PokemonGoBot(object):
         if self.config.evolve_all:
             # Run evolve all once. Flip the bit.
             print('[#] Attempting to evolve all pokemons ...')
-            self.config.evolve_all = False
             worker = EvolveAllWorker(self)
             worker.work()
+            self.config.evolve_all = []
 
         self._filter_ignored_pokemons(cell)
 
@@ -54,16 +55,18 @@ class PokemonGoBot(object):
             cell['catchable_pokemons'].sort(
                 key=
                 lambda x: distance(self.position[0], self.position[1], x['latitude'], x['longitude']))
-            for pokemon in cell['catchable_pokemons']:
-                with open('web/catchable-%s.json' %
-                          (self.config.username), 'w') as outfile:
-                    json.dump(pokemon, outfile)
 
-                if self.catch_pokemon(pokemon) == PokemonCatchWorker.NO_POKEBALLS:
-                    break
-                with open('web/catchable-%s.json' %
-                          (self.config.username), 'w') as outfile:
-                    json.dump({}, outfile)
+            user_web_catchable = 'web/catchable-%s.json' % (self.config.username)
+            if os.path.isfile(user_web_catchable): # only write to file if it exists
+                for pokemon in cell['catchable_pokemons']:
+                    with open(user_web_catchable, 'w') as outfile:
+                        json.dump(pokemon, outfile)
+
+                    if self.catch_pokemon(pokemon) == PokemonCatchWorker.NO_POKEBALLS:
+                        break
+                    with open(user_web_catchable, 'w') as outfile:
+                        json.dump({}, outfile)
+
         if (self.config.mode == "all" or self.config.mode == "poke"
             ) and 'wild_pokemons' in cell and len(cell['wild_pokemons']) > 0:
             # Sort all by distance from current pos- eventually this should
@@ -121,7 +124,7 @@ class PokemonGoBot(object):
         # check if the release_config file exists
         try:
             with open('release_config.json') as file:
-               pass
+                pass
         except:
             # the file does not exist, warn the user and exit.
             logger.log('[#] IMPORTANT: Rename and configure release_config.json.example for your Pokemon release logic first!', 'red')
@@ -232,9 +235,11 @@ class PokemonGoBot(object):
         inventory_req = self.api.call()
         inventory_dict = inventory_req['responses']['GET_INVENTORY'][
             'inventory_delta']['inventory_items']
-        with open('web/inventory-%s.json' %
-                  (self.config.username), 'w') as outfile:
-            json.dump(inventory_dict, outfile)
+
+        user_web_inventory = 'web/inventory-%s.json' % (self.config.username)
+        if os.path.isfile(user_web_inventory):
+            with open(user_web_inventory, 'w') as outfile:
+                json.dump(inventory_dict, outfile)
 
         # get player balls stock
         # ----------------------
@@ -243,21 +248,18 @@ class PokemonGoBot(object):
         for item in inventory_dict:
             try:
                 # print(item['inventory_item_data']['item'])
-                if item['inventory_item_data']['item'][
-                        'item_id'] == Item.ITEM_POKE_BALL.value:
-                    # print('Poke Ball count: ' + str(item['inventory_item_data']['item']['count']))
-                    balls_stock[1] = item[
-                        'inventory_item_data']['item']['count']
-                if item['inventory_item_data']['item'][
-                        'item_id'] == Item.ITEM_GREAT_BALL.value:
-                    # print('Great Ball count: ' + str(item['inventory_item_data']['item']['count']))
-                    balls_stock[2] = item[
-                        'inventory_item_data']['item']['count']
-                if item['inventory_item_data']['item'][
-                        'item_id'] == Item.ITEM_ULTRA_BALL.value:
-                    # print('Ultra Ball count: ' + str(item['inventory_item_data']['item']['count']))
-                    balls_stock[3] = item[
-                        'inventory_item_data']['item']['count']
+                item_id = item['inventory_item_data']['item']['item_id']
+                item_count = item['inventory_item_data']['item']['count']
+
+                if item_id == Item.ITEM_POKE_BALL.value:
+                    # print('Poke Ball count: ' + str(item_count))
+                    balls_stock[1] = item_count
+                if item_id == Item.ITEM_GREAT_BALL.value:
+                    # print('Great Ball count: ' + str(item_count))
+                    balls_stock[2] = item_count
+                if item_id == Item.ITEM_ULTRA_BALL.value:
+                    # print('Ultra Ball count: ' + str(item_count))
+                    balls_stock[3] = item_count
             except:
                 continue
         return balls_stock
@@ -283,15 +285,23 @@ class PokemonGoBot(object):
     def _set_starting_position(self):
 
         if self.config.test:
+            # TODO: Add unit tests
             return
 
         if self.config.location:
             try:
                 location_str = str(self.config.location)
-                start_coordinate = [x.strip() for x in location_str.split(',')]
-                self.position = (float(start_coordinate[0]), float(start_coordinate[0]), 0.0)
+                location = (self._get_pos_by_name(location_str.replace(" ", "")))
+                self.position = location
                 self.api.set_position(*self.position)
+                logger.log('')
+                logger.log(u'[x] Address found: {}'.format(self.config.location.decode(
+                    'utf-8')))
+                logger.log('[x] Position in-game set as: {}'.format(self.position))
+                logger.log('')
+                return
             except:
+                logger.log('[x] The location given using -l could not be parsed. Checking for a cached location.')
                 pass
 
         if self.config.location_cache and not self.config.location:
@@ -322,16 +332,6 @@ class PokemonGoBot(object):
                         "No cached Location. Please specify initial location.")
                 else:
                     pass
-
-        #
-        # this will fail if the location.json isn't there or not valid.
-        # Still runs if location is set.
-        self.position = self._get_pos_by_name(self.config.location)
-        self.api.set_position(*self.position)
-        logger.log('')
-        logger.log(u'[x] Address found: {}'.format(self.config.location))
-        logger.log('[x] Position in-game set as: {}'.format(self.position))
-        logger.log('')
 
     def _get_pos_by_name(self, location_name):
         # Check if the given location is already a coordinate.
